ft_printf normify

+++ lib/libft/ft_dtoa/ft_dtoa.c
@@ -6,7 +6,7 @@
 /*   By: ohakola <ohakola@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/17 18:19:22 by ohakola           #+#    #+#             */
-/*   Updated: 2021/05/04 16:19:52 by ohakola          ###   ########.fr       */
+/*   Updated: 2021/05/05 11:43:33 by ohakola          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -67,7 +67,7 @@ static void	set_dragon4_params(t_dragon4_params *dragon, t_dtoa_params dtoa,
 {
 	t_float_dissector	fd;
 
-	fd.f = dtoa.value;
+	fd.f = (double)dtoa.value;
 	dragon->no_trailing_zeros = dtoa.g_mode;
 	dragon->hashtag = dtoa.hashtag;
 	if (fd.b.exp != 0)
@@ -102,7 +102,7 @@ char	*ft_dtoa(t_dtoa_params params)
 	t_dragon4_params	dragon;
 	t_float_dissector	fd;
 
-	fd.f = params.value;
+	fd.f = (double)params.value;
 	if (fd.b.sign == 1)
 		*buf = '-';
 	if (fd.b.exp == 0x7FF)
@@ -131,7 +131,7 @@ int	ft_dtoa_buf(t_dtoa_params params, char *buf, int buf_size)
 	t_float_dissector	fd;
 	uint32_t			inf_nan_sign;
 
-	fd.f = params.value;
+	fd.f = (double)params.value;
 	if (fd.b.sign == 1)
 		*buf = '-';
 	inf_nan_sign = (fd.b.sign == 1 && fd.b.fraction == 0);
